+++ app/Domain/Mail/Mailer.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Domain\Mail;
 
-use App\Mails\Addresses;
 use Symfony\Component\Mime\Address;
 
 interface Mailer
