+++ config/app.php
@@ -204,7 +204,7 @@ return [
         App\ProjectAanvraag\ProjectAanvraagServiceProvider::class,
         App\Keycloak\KeycloakServiceProvider::class,
         App\Search\SearchServiceProvider::class,
-        App\Mails\MailServiceProvider::class,
+        \App\MailJet\MailjetServiceProvider::class,
         \App\Notifications\NotificationsProvider::class,
     ],
