+++ tests/Mails/MailManagerTest.php
@@ -13,9 +13,9 @@ use App\Domain\Integrations\IntegrationPartnerStatus;
 use App\Domain\Integrations\IntegrationStatus;
 use App\Domain\Integrations\IntegrationType;
 use App\Domain\Integrations\Repositories\IntegrationRepository;
+use App\Domain\Mail\Addresses;
 use App\Domain\Mail\Mailer;
-use App\Mails\Addresses;
-use App\Mails\MailManager;
+use App\MailJet\MailManager;
 use PHPUnit\Framework\MockObject\MockObject;
 use Ramsey\Uuid\Uuid;
 use Symfony\Component\Mime\Address;
